# [do-not-merge] db: add NextPrefix optimization