--- a/error_iter.go
+++ b/error_iter.go
@@ -43,7 +43,7 @@ func (c *errorIter) Next() (*InternalKey, []byte) {
 	return nil, nil
 }
 
-func (c *errorIter) NextPrefix() (*InternalKey, []byte) {
+func (c *errorIter) NextPrefix(int) (*InternalKey, []byte) {
 	return nil, nil
 }