--- a/internal/arenaskl/iterator.go
+++ b/internal/arenaskl/iterator.go
@@ -194,6 +194,11 @@ func (it *Iterator) Next() (*base.InternalKey, []byte) {
 	return &it.key, it.value()
 }
 
+// NextPrefix ...
+func (it *Iterator) NextPrefix() (*base.InternalKey, []byte) {
+	return it.Next()
+}
+
 // Prev moves to the previous position. Returns the key and value if the
 // iterator is pointing at a valid entry, and (nil, nil) otherwise.
 func (it *Iterator) Prev() (*base.InternalKey, []byte) {
@@ -207,6 +212,11 @@ func (it *Iterator) Prev() (*base.InternalKey, []byte) {
 		return nil, nil
 	}
 	return &it.key, it.value()
+}
+
+// PrevPrefix ...
+func (it *Iterator) PrevPrefix() (*base.InternalKey, []byte) {
+	return it.Prev()
 }
 
 // value returns the value at the current position.