--- a/iterator.go
+++ b/iterator.go
@@ -239,7 +239,7 @@ func (i *Iterator) findNextEntry(limit []byte) {
 
 		switch key.Kind() {
 		case InternalKeyKindDelete, InternalKeyKindSingleDelete:
-			i.nextUserKey()
+			i.nextUserKey(false)
 			continue
 
 		case InternalKeyKindSet, InternalKeyKindSetWithDelete:
@@ -261,7 +261,7 @@ func (i *Iterator) findNextEntry(limit []byte) {
 				if i.err == nil && needDelete {
 					i.iterValidityState = IterExhausted
 					if i.pos != iterPosNext {
-						i.nextUserKey()
+						i.nextUserKey(false)
 					}
 					if i.closeValueCloser() == nil {
 						i.pos = iterPosCurForward
@@ -291,7 +291,7 @@ func (i *Iterator) closeValueCloser() error {
 	return i.err
 }
 
-func (i *Iterator) nextUserKey() {
+func (i *Iterator) nextUserKey(skipPrefix bool) {
 	if i.iterKey == nil {
 		return
 	}
@@ -301,7 +301,11 @@ func (i *Iterator) nextUserKey() {
 		i.key = i.keyBuf
 	}
 	for {
-		i.iterKey, i.iterValue = i.iter.Next()
+		if skipPrefix {
+			i.iterKey, i.iterValue = i.iter.NextPrefix()
+		} else {
+			i.iterKey, i.iterValue = i.iter.Next()
+		}
 		i.stats.ForwardStepCount[InternalIterCall]++
 		if done || i.iterKey == nil {
 			break
@@ -931,8 +935,19 @@ func (i *Iterator) Next() bool {
 	return i.NextWithLimit(nil) == IterValid
 }
 
+// NextPrefix moves the iterator to the next key/value pair with a different
+// prefix than the current key.
+func (i *Iterator) NextPrefix() bool {
+	return i.nextWithLimit(nil, true) == IterValid
+}
+
 // NextWithLimit ...
 func (i *Iterator) NextWithLimit(limit []byte) IterValidityState {
+	return i.nextWithLimit(limit, false)
+}
+
+// nextWithLimit ...
+func (i *Iterator) nextWithLimit(limit []byte, skipPrefix bool) IterValidityState {
 	i.stats.ForwardStepCount[InterfaceCall]++
 	if limit != nil && i.hasPrefix {
 		i.err = errors.New("cannot use limit with prefix iteration")
@@ -945,7 +960,7 @@ func (i *Iterator) NextWithLimit(limit []byte) IterValidityState {
 	i.lastPositioningOp = unknownLastPositionOp
 	switch i.pos {
 	case iterPosCurForward:
-		i.nextUserKey()
+		i.nextUserKey(skipPrefix)
 	case iterPosCurForwardPaused:
 		// Already at the right place.
 	case iterPosCurReverse:
@@ -974,7 +989,7 @@ func (i *Iterator) NextWithLimit(limit []byte) IterValidityState {
 			i.iterValidityState = IterExhausted
 			return i.iterValidityState
 		}
-		i.nextUserKey()
+		i.nextUserKey(false)
 	case iterPosPrev:
 		// The underlying iterator is pointed to the previous key (this can
 		// only happen when switching iteration directions). We set
@@ -993,9 +1008,9 @@ func (i *Iterator) NextWithLimit(limit []byte) IterValidityState {
 				i.stats.ForwardSeekCount[InternalIterCall]++
 			}
 		} else {
-			i.nextUserKey()
+			i.nextUserKey(false)
 		}
-		i.nextUserKey()
+		i.nextUserKey(false)
 	case iterPosNext:
 		// Already at the right place.
 	}
@@ -1007,6 +1022,12 @@ func (i *Iterator) NextWithLimit(limit []byte) IterValidityState {
 // Prev moves the iterator to the previous key/value pair. Returns true if the
 // iterator is pointing at a valid entry and false otherwise.
 func (i *Iterator) Prev() bool {
+	return i.PrevWithLimit(nil) == IterValid
+}
+
+// PrevPrefix moves the iterator to the previous key/value pair with a different
+// prefix than the current key.
+func (i *Iterator) PrevPrefix() bool {
 	return i.PrevWithLimit(nil) == IterValid
 }