--- a/sstable/block.go
+++ b/sstable/block.go
@@ -324,7 +324,7 @@ func (i *blockIter) resetForReuse() blockIter {
 	}
 }
 
-func (i *blockIter) readEntry() {
+func (i *blockIter) readEntry() (shared uint32) {
 	ptr := unsafe.Pointer(uintptr(i.ptr) + uintptr(i.offset))
 
 	// This is an ugly performance hack. Reading entries from blocks is one of
@@ -335,7 +335,6 @@ func (i *blockIter) readEntry() {
 	//
 	// TODO(peter): remove this hack if go:inline is ever supported.
 
-	var shared uint32
 	if a := *((*uint8)(ptr)); a < 128 {
 		shared = uint32(a)
 		ptr = unsafe.Pointer(uintptr(ptr) + 1)
@@ -406,6 +405,7 @@ func (i *blockIter) readEntry() {
 	ptr = unsafe.Pointer(uintptr(ptr) + uintptr(unshared))
 	i.val = getBytes(ptr, int(value))
 	i.nextOffset = int32(uintptr(ptr)-uintptr(i.ptr)) + int32(value)
+	return shared
 }
 
 func (i *blockIter) readFirstKey() error {
@@ -834,8 +834,38 @@ func (i *blockIter) Next() (*InternalKey, []byte) {
 }
 
 // NextPrefix ...
-func (i *blockIter) NextPrefix() (*InternalKey, []byte) {
-	return i.Next()
+func (i *blockIter) NextPrefix(currentPrefixLen int) (*InternalKey, []byte) {
+	if len(i.cachedBuf) > 0 {
+		// We're switching from reverse iteration to forward iteration.
+		return i.Next()
+	}
+
+	for {
+		i.offset = i.nextOffset
+		if !i.Valid() {
+			return nil, nil
+		}
+		shared := i.readEntry()
+		if shared > uint32(currentPrefixLen) {
+			// This key must have the same prefix as the key that the user last
+			// saw, so skip it.
+			continue
+		}
+
+		// Manually inlined version of i.decodeInternalKey(i.key).
+		if n := len(i.key) - 8; n >= 0 {
+			i.ikey.Trailer = binary.LittleEndian.Uint64(i.key[n:])
+			i.ikey.UserKey = i.key[:n:n]
+			if i.globalSeqNum != 0 {
+				i.ikey.SetSeqNum(i.globalSeqNum)
+			}
+		} else {
+			i.ikey.Trailer = uint64(InternalKeyKindInvalid)
+			i.ikey.UserKey = nil
+		}
+		break
+	}
+	return &i.ikey, i.val
 }
 
 // Prev implements internalIterator.Prev, as documented in the pebble