This is a very crude proof of concept of an idea to mitigate the problem of
MVCC garbage. The Iterator interface is expanded with a NextPrefix method.
When a client no longer cares about the prefix currently positioned under
the iterator, it may call NextPrefix to advance to the next distinct prefix.
In this proof of concept, NextPrefix does not guarantee it will surface a new
prefix, although we could if we buffer the current prefix.

Internally, NextPrefix behaves exactly like a Next, except that the internal
iterator positioned at the top of the merging iterator's heap is advanced with
NextPrefix instead of Next. This gives that iterator an opportunity to use its
own context to skip over keys.

This change adds one such optimization, gated behind a `PrefixSkipIndexing`
flag. If set, Pebble configures an internally-managed block property
collector. As an sstable is constructed, this block property collector watches
for changes in key prefixes. For every block and index block, it records
whether the key prefix changed within the block or index block's span. If the
prefix did change, the collector elides any block property. If the prefix did
not change, the collector adds a 1-byte property singaling that the block or
index block does not contain any changes in prefix from the perspective of
forward iteration.

When NextPrefix is called on a sstable Reader and the current block is
exchausted, the Reader extracts the next block's property. If it exists,
indicating the next block did not observe any prefix change during
construction, the next block is irrelevant and may be skipped.

This optimization is tailored towards forward iteration. Providing the same
optimization for backward iteration is trickier, because sstables are
constructed in forward iteration and the prefix change indicators are
generated from that forward iteration. However, backward iteration may still
skip a block if both its block property and the current exhausted block's
property indicate that neither contains a prefix change. This isn't
implemented in this proof of concept.

I also made an additional optimization with NextPrefix that helps avoid key
comparisons for versions of the same key within a block. The NextPrefix
InternalIterator interface propagates the length of the current prefix, which
can be used by the block iterator NextPrefix to skip over keys so long as the
number of bytes shared in prefix sharing is greater than the current prefix
length.

I ran a benchmark on my gceworker of a forward scan through a keyspace with
varying numbers of duplicate keys. All keys had 10-byte values.:

```
name                          old time/op    new time/op    delta
NextPrefix/versions=1-24         142µs ± 5%     150µs ± 4%   +5.86%  (p=0.032 n=5+5)
NextPrefix/versions=10-24       1.87ms ± 1%    0.40ms ± 0%  -78.46%  (p=0.008 n=5+5)
NextPrefix/versions=100-24      12.6ms ± 3%     2.6ms ± 2%  -79.31%  (p=0.008 n=5+5)
NextPrefix/versions=1000-24      147ms ± 2%       6ms ± 2%  -95.88%  (p=0.008 n=5+5)
NextPrefix/versions=10000-24     1.48s ± 1%     0.02s ± 7%  -98.74%  (p=0.008 n=5+5)

name                          old alloc/op   new alloc/op   delta
NextPrefix/versions=1-24         10.0B ± 0%     10.0B ± 0%     ~     (all equal)
NextPrefix/versions=10-24        38.0B ± 0%     14.4B ± 4%  -62.11%  (p=0.000 n=4+5)
NextPrefix/versions=100-24        141B ±41%       37B ± 3%  -73.83%  (p=0.008 n=5+5)
NextPrefix/versions=1000-24     3.95kB ±16%    0.15kB ± 3%  -96.19%  (p=0.008 n=5+5)
NextPrefix/versions=10000-24     169kB ± 2%       4kB ±12%  -97.82%  (p=0.008 n=5+5)

name                          old allocs/op  new allocs/op  delta
NextPrefix/versions=1-24          1.00 ± 0%      1.00 ± 0%     ~     (all equal)
NextPrefix/versions=10-24         1.00 ± 0%      1.00 ± 0%     ~     (all equal)
NextPrefix/versions=100-24        1.00 ± 0%      1.00 ± 0%     ~     (all equal)
NextPrefix/versions=1000-24       34.8 ±18%       2.0 ± 0%  -94.25%  (p=0.008 n=5+5)
NextPrefix/versions=10000-24     1.74k ± 2%     0.05k ± 7%  -96.99%  (p=0.008 n=5+5)
```

I suspect the overhead for 1-version per key can be reduced by only attempting
this optimization if we've already seen the same prefix multiple times.