@sumeerbhola 2021-11-22
Nice!

Regarding the following
```
NextPrefix/versions=10-24       1.87ms ± 1%    0.40ms ± 0%  -78.46%  (p=0.008 n=5+5)
NextPrefix/versions=100-24      12.6ms ± 3%     2.6ms ± 2%  -79.31%  (p=0.008 n=5+5)
```
with ~20 bytes for each key-value pair, with 100 versions we have ~2000 bytes, so the block property optimization would not help with 32KB blocks. Are all these gains coming from avoiding the key prefix comparison?