@sumeerbhola 2021-11-23
Can you pull out the key prefix comparison optimization into a separate PR. I think the `NextPrefix` method plus that optimization is clean enough to merge.