@jbowens 2021-12-08
Work-in-progress productionizing of the prefix-compression optimization is happening in #1387. I'm going to close this out as it's not intended to be merged, and we can consult the closed pull request if we ever consider productionizing the block-property collector optimization.